API/requirements.txt
fastapi==0.92.0
uvicorn==0.20.0
Pillow==9.4.0
python-multipart==0.0.6